Tighten up enum param comparisons
Prior to this change, when two enums were param compared using ==
or !=, we would compare their integer values rather than seeing
whether or not they actually are the same enum.  This seems like
it's wrong.  With this patch, if the two enums are of different
enum types, we return false (they're not the same); if they're
of the same type, we proceed as before.

For inequalities between enums (<, >, <=, >=), everything is as
before.  It seems that these operators only make sense for enums
on the integer values.