Only apply concrete where clause evaluation to instantiated functions
We seem to have some concrete functions with where clauses whose
presence breaks the evaluation of where clauses for all concrete
functions.  But if we limit it just to the concrete functions
that represent instantiations of generic functions, as in this
change, I get a good result from my futures (the generic ones
that is) and also get promotion cases working that weren't
without this change.

My hunch is that if we put an else clause on this new conditional, we
could easily find the concrete functions that have bogus where
clauses.