Fix a .future so it fails on linux64
This .future assumed that "int64_t" was equivalent to "long long" on 64 bit
platforms, which was true on OSX.  On a Linux64 testing system, "int64_t" was
equivalent to "long".  The Chapel types "c_long" and "c_longlong" both turn
into "int64_t" on these platforms.

Add another case to the test that passes a "c_longlong" by reference to an
extern function expecting a "long long*".  This way one of the cases will fail
whichever type "int64_t" defaults to.