Turn chpl__unalias into a returning function
For array-views, we want chpl__unalias to create a type different from
the formal. This is more easily accomplished by turning chpl__unalias
into a returning function instead of one that modifies the formal
in-place.

The compiler modifications to support this change are fairly simple. We
now allow for the reduction of initCopy/autoCopy if the types differ and
there's an available chpl__unalias function. We then create new AST to
capture the result of chpl__unalias, and move the result into the
correct positions.

chpl__unalias functions now have to return something. Existing
chpl__unalias functions have been modified to include some path that
returns the formal without modifications. Otherwise, the chpl__unalias
functions for domains and arrays now invoke chpl__initCopy implicitly to
perform the deep-copy when necessary (e.g., for aliases).