Mark leader promotion-wrapper iterators as generic
Add FLAG_GENERIC to the leader iterators that are built for promotion wrappers.
Not doing so causes compilation errors when trying to evaluate where clauses
for non-generic functions.

I have no idea how this hasn't caused other errors in the past (like failing to
resolve the wrapper leader iterator) but apparently it hasn't and I can't
construct a test that fails.