Stop ignoring where-clauses for non-generic functions
Evaluate and respect where-clauses for non-generic functions. This is a
follow-on to #5275.

Previously the language definition stated that "The optional where-clause is
only applicable if the function is generic". The compiler didn't error if a
where-clause was attached to a non-generic function, it just never evaluated
it, effectively pretending the where-clause wasn't there.

You could write something like:

    proc foo(x: int) where bar() { // note that there is no bar()
      writeln("In foo()");
    }
    foo(1);

and the compiler would ignore the bogus `where bar()` and happily resolve foo()

Originally, we didn't see a use-case for where-clauses on non-generic
functions, which is why the spec disallowed it. We imagined that where-clauses
would only compute on aspects of the functions arguments, so if the function
was non-generic, there'd be no need for a where-clause. We've since come to
realize that it's more orthogonal to support where-clauses on non-generics and
that there are real use-cases for it such as selecting a function
implementation based off a config param:

    config param matrixLayout = row_major;

    proc matrixOp(A, B) where matrixLayout == row_major { /* row-major impl */ }
    proc matrixOp(A, B) where matrixLayout == col_major { /* column-major impl */ }