Fix some tests with bad where-clauses
These tests did something like:

    proc foo(a: int) where a == 42 { writeln("foo1"); }
    proc foo(a: int)               { writeln("foo2"); }

which is invalid since `a` wasn't marked param. However, foo was non-generic so
we didn't evaluate the where-clause, and thus the compiler ignored the error.

This test happened to work because even though we didn't evaluate the
where-clause, simply having a where-clause made the function "more specific" so
it was picked in-favor of the where-less function.