Add a test throw destroys local records
This is a simple test and it is implemented oddly because we don't have
catch blocks yet...

The intent here is to just make sure that a function that throws still
cleans up local variables appropriately on the throwing path. That means,
in particular, calling a record destructor, which is what this test
exercises.