Edits to CHIP 10 on deinit() ----------------------------
Clarified that a compiler-generated deinit() is equivalent
to an empty user-defined deinit().

Added the rationale for current rules.

Minor style changes.

More changes needed
-------------------

While reading the CHIP, I realized that

* We need to define deinitialization at least for arrays.

* Deinitialization makes sense also for domains,
given that the domain map author may want to perform some actions
(like freeing resources) when the domain is deleted.

* To dot all the Is, we should at some point look through
all Chapel types and see which ones need to have deinitialization
defined or be not a NOP.

  As an intermediate step we could say things like "defining the actions
that happen when a domain is deinitialized is future work."

* The current CHIP imposes deinitialization order on record fields
and hints on that requirement for array/domain fields. Meanwhile
Brad in his email on 1/26 suggested, and we all agreed,
reverse-declaration order of deinitialization for **all** fields.
Should we change the CHIP accordingly?