Change unset of 'CHPL_UNWIND' to warnings + optional exit in start_test
An unanticipated impact of my change to unset CHPL_UNWIND in
start_test (#5184) is that if you've got CHPL_UNWIND set, build the
runtime, and then run start_test, you may not have the correct runtime
built and up-to-date since we specialize the runtime builds based on
the CHPL_UNWIND setting.  Thus, you could get an error saying that the
proper runtime configuration isn't built (even though you thought you
just built it), or you could inadvertantly test a stale runtime that
you had lying around.

This change backs out the unsetting of CHPL_UNWIND and changes it to
a warning being generated if CHPL_UNWIND is set to something other
than 'none' when testing is run.  It also adds an environment variable
CHPL_TEST_HALT_IF_UNWIND_ENABLED which, if set, will cause start_test
to exit if CHPL_UNWIND is set as a means of catching such cases prior
to potentially running the entire test suite.