Add tests for the date and datetime types
These tests are based on a python unit test for the python datetime module.

Fixed a few problems in date and datetime methods that these tests caught.

Added a variable that uses datetime.today() at the top of the DateTime module
to work around issue #5267.  This way programs that use the DateTime module
don't need to work around that issue themselves.