Add tests demonstrating surprise w.r.t. array fields and constructors
This pair of tests demonstrates that the compiler-generated
constructor is arguably surprising w.r.t. array fields as compared to
a similar constructor that a user might write in an attempt to model
what the compiler-generated constructor would look like.

Specifically, if the array's field is a default generic array, the
current compiler-generated constructor only permits default
rectangular arrays to be passed in for that array argument, blocking
other domain maps, like the block-distributed array in this example.
By contrast, in the user constructor, the use of an a formal type that
resembles a default rectangular type behaves similarly to other
function call contexts where it asserts something about the array's
size/shape/rank but permits other domain maps to be passed in.

One piece of why these cases are treated differently seems to be in
normalize.cpp where the fixup_array_formals() call is filtered so as
not to apply to the default constructor or type constructor.