Establish bulk-transfer interface for ArrayViews
ArrayViews will support a '_getViewDom()' method which returns a
first-class domain (a _domain record) that may or may not be
distributed. This domain must match the rank of the original BaseArr
class.

The tricky case is with nested rank-changes. I have modified the
'chpl_rankChangeConvertDom' method to accept a tuple of ranges and to
return a first-class domain. If there are nested rank-changes, we
essentially call `chpl_rankChangeConvertDom` and pass the dimensions of
the resulting domain down to the nested rank-change for translation.

Three utility functions have been added:
chpl__isArrayView(arr:BaseArr) param : bool
  Returns true if 'arr' is an ArrayView
chpl__getViewDom(arr:[]) : domain
  Returns a view-domain for the given array. If the array is not an
  array-view, it returns arr.domain
chpl__getActualArray(arr) : BaseArr
  'arr' can either be a first-class array or a BaseArr. If it is an
  array-view, returns the original BaseArr. Otherwise returns the
  top-level BaseArr of 'arr'

Minor fix to [].equals([]) to not check dimensions if the arrays are not
rectangular. Fixes a problem with associative arrays.