Revise ZMQ/weather prediff from PR #5321 to hide fewer test errors.
By coincidence, the day we merged PR #5321, an unrelated PR #5303
went in the same day; and it made a change which should have broke
the ZMQ/weather test. A new Warning message is issued when chpl
compiles weather.chpl, completely unrelated to the dropped ZMQ
messages that prompted PR #5321.  Normally, this new chpl Warning
message would show up in the test results, and the test harness
would flag an error.

Unfortunately, the new weather.prediff from PR #5321 hides this
new chpl Warning from the test harness, and ZMQ/weather test
passes when it should fail.

This change rewrites ZMQ/weather.prediff so it ONLY hides the
kind of test error that PR #5321 was meant for.

A fix for the new chpl Warning message from ZMQ/weather.chpl is
in a separate change, not discussed here.