Adding a few more cases to Sung's array memory test
Specifically, previously the pass/return reindex() functions had
changed pretty dramatically from the slice cases when we moved
the reindex into the actual argument.  Here, I create versions
of slice that do the slicing in the actual and also versions of
the reindex that do the reindexing within the function for
symmetry.  This works well on master, but I believe causes
problems on the array views branches.