Extension of CallInfo to support initializers
When an initializer is called, if the type being initialized is generic, then
the type of the actual will not be determined until we have resolved phase 1 of
the body of the best match.  Normally, CallInfo creation would reject such calls
immediately.  This change convinces CallInfo to let it pass for now.