Rearrange instantiateSignature
I have created three new functions from the contents of instantiateSignature:
determineRootFunc, determineAllSubs, and instantiateFunction.  These take
key bits of code that will be used in my version of the same function for
generic initializers, so that the implementation of both will be more
maintainable going forward.  At least, that is my hope.

I have not gotten all the way through the function just yet, this is saving
my work.