Remove "generic" arg from evaluateWhereClauses
This was added as a workaround in #5275, but it's no longer needed since we're
no longer evaluating where-clauses after they've been instantiated.

Note that ideally we'd move evaluateWhereClauses to functionResolution since
that's the only pace it's used now, but right now evaluateWhereClauses
interacts with a global whereStack and checkInfiniteWhereInstantiation() in
generics.cpp, so it's a pretty big hassle to move it.