Refine _getViewDom functionality for ArrayViews
This commit implements the following methods on ArrayViewReindex:
- chpl_reindexConvertDom
- _getViewDom
- _viewHelper

These methods allow for reindexes to be bulk-transferred. In support of
this change, the following functions were modified:
- _containsRankChange --> _containsRCRE
- _getRankChangeView  --> _getRCREView

Here 'RCRE' stands for "RankChange/Reindex". These two kinds of
ArrayViews need to be able to handle RankChanges and Reindexes further
down a view stack. We don't care about slices further down the stack
because the dimensions at the top are already sliced correctly.

This commit also implements a number of bulk-transfer interface methods
on ArrayViewReindex, much like the other ArrayViews.

The hpl.hpcc2012 test was modified to avoid compile-time errors, but
instead now encounters execution-time failures.