Added future on passing subclass to superclass ref formal
When a [const] ref formal has the type C, the actual also must
have the type C. When the program passes a D actual, where D
is a subclass of C, the compiler fails verification or generates
invalid C code.

This is based on a bug posted by a user to chapel-users list today.