Stop treating reindexing errors on gasnet as "correct"
This test represents a case that the block distribution was
unable to reindex properly, so we locked a halt message into
the .good file instead.  With the array view work, it is
better behaved and so the same .good file can be used for
local and gasnet testing, and the prediff removed.