Arrays always autoCopy when returned.
The function ``insertAutoCopyForArrayReturn`` inserts an autoCopy call
when it an _array record is assigned to the function's returnSymbol
(unless the function or RHS expression has certain properties). If the
autoCopy is not resolved, it will be resolved here.

A change was made in callDestructors to avoid removing the autoCopy if
it saw that the LHS was the return-value-variable.

FLAG_NO_COPY_RETURN was added as a quick way to mark internal functions
that shouldn't deep-copy when returning an array.

As a result, there is no need for specialized versions of chpl__initCopy
and chpl__autoCopy for ArrayViews.