Skip this test for CHPL_ATOMICS=locks
I tried for a few minutes to run the test and link with the
Chapel runtime but I couldn't quickly figure out how to do so.
Therefore, for now, this test assumes that the relevant
functionality is in header files, and it skips CHPL_ATOMICS
which is the case where we know it is not.