Make var-args I/O functions use const intent
My improved const-checking doesn't yet work correctly with tuples (as
used in varargs functions like writeln). At the same time, I'm not sure
that it's reasonable to get tuple const-checking working in the short
term, as it might require the compiler to actually create different tuple
types depending on whether or not the elements are const.

So, as a work-around, vararg functions that will be passed any
const variables need to be marked with const.