Get things working for clang/clang++ + clean up
The trick here was to not let the launcher fall back to the default
LD = ld linker.  Compiling with clang/clang++ in my machine yet
linking with 'ld' seems to result in warnings of the form:

ld: warning: object file (/Users/bradc/chapel/lib/darwin/gnu/loc-flat/comm-gasnet/tasks-qthreads/launch-amudprun/tmr-generic/unwind-system/mem-jemalloc/atomics-intrinsics/wide-struct/libchpllaunch.a(launch-amudprun.o)) was built for newer OSX version (10.11) than being linked (10.10)

The trick was to make the launcher set LD to CC.  Once that got things
working, I cleaned up the other paths.  At present, we assume that the
Makefile.comm-* file sets the linker for the _real binary and that the
Makefile.launcher file sets it for the launcher binary.