Set LD on other branches within GASNet Makefile
On my previous edit, I lost track of the fact that the set of LD
was within a conditional and that I was failing to set LD along
the other branches in the conditional.  Recall that on Crays,
for example, we don't use GASNet's settings because it causes
paths to be hardcoded more than we'd like for portability of our
module.  Here, I set it to CXX as we do in comm-none for the same
reason.