Declare ISx's arrays over an anonymous range instead of a remote domain
Declaring an array over a remote domain results in an on-statement onto
that domain's locale in order to increment the reference count and
manipulate the list of arrays declared over the domain. This can result
in a potentially large number of on-statements on Locale0, which can
hurt performance and hold up the exchange's barrier. Slicing a remote
domain also introduces some communication.