Move contents of renameInstantiatedType into a series of methods on TypeSymbols
I reuse and modify some of this code for my generic initializers branch.
Really, it makes the most sense for this operation to be a method on the
TypeSymbol rather than an outside function.  This will improve code
maintainability for my initializers implementation.

This commit starts that process by moving the renaming that always occurs on
an instantiation to a private method, renameInstantiatedStart, and calling that
method from a public method, renameInstantiatedMulti, which will handle
renaming when multiple substitutions are performed at a time.

Verified that the generated log output and C code do not differ beyond normal
variability (i.e. our compiler doesn't guarantee the ordering of certain symbols
from run to run).