Last of the rearrangements of the body
I split the innards of the loop into the part that was loop dependent and the
part that wasn't, moving the inner workings that would apply to most into the
method renameInstantiatedIndividual and leaving the rest in the loop body of
renameInstantiatedMulti

Verified that the generated C code and log output was unchanged, modulo normal
variability, by diff'ing the log output of hello.chpl after resolution and the
C code and looking through the results.  Since our library code encounters a lot
of the edge cases (and I caught an error with this strategy), I believe this is
sufficient to guarantee no change in behavior, after I run a full paratest.