Get HPL HPCC2012 test passing in ArrayViews world
This test suffered from two problems:
1) dsiLocalSlice1 returned a reindex of a rank-change or slice. While
the reindex may have been returned without a copy or autodestroy, the
inner ArrayView would be destroyed. This could lead to segfaults or
array OOB errors. The solution is to alias the inner ArrayView and place
a 'no auto destroy' pragma on it. Long-term, I think a general-purpose
non-stacking ArrayView class could resolve this issue.

2) The ddata calculations were no longer correct. In the closed-form
world, the DefaultRectangular was set up such that indexing was less
costly and require no translation. That is not the case in ArrayViews.
For now the easiest solution was to remove the ddata accesses and use
normal 'this[]' accesses on the ArrayViews.