Retire futures that pass under array view testing
The following futures can be retired as part of the array view work:

* test/distributions/bradc/blkCyc/blkCycSliceOOB.chpl
* test/distributions/dm/sx8.chpl
* test/optimizations/rafa/rankChange.future

The second suggested that when the test passed, s8.chpl and s9.chpl
should be changed to include an additional line, so I've done that
and removed sx8.*.

The third suggested that when the test passed, disableAliasedBulkTransfer
could be removed.  I haven't done that yet because it's a bigger
change and I want to be sure that things will continue to work when
we have distributed rank change domains.