Squash some wide pointer overhead for arrays
IWR will now pretend that an _array record is actually the _instance
class inside of that record. This allows for the _instance field to be
inferred to be local (when being accessed), which improves performance
for array indexing.

With the ArrayViews work in, the wide-return optimization is no longer
very useful. I have removed it because:
- it is complex and ugly
- it has been the source of bugs
- it will hopefully be re-implemented in a more general way in the
  future, wihout all the iterative stuff