Work around insufficiently generic compiler-provided constructors
This works around issue #5289 which causes problems for this test on
the array views branch by creating a clone of the test that uses a
user constructor to work around the limitation and futurizing the
original test.  When the original test starts working again, we can
retire the workaround (unless we think it adds value).