Remove the single-yield iterator inlining "optimization"
The inlineSingleYieldIterator optimization was originally added as part of some
prototype GPU code with af0e77e. Most of the GPU code was later removed in
60c2580, but inlineSingleYieldIterator remained.

The only thing this optimization does is permit zippered inlining of iters that
only yield a single value. i.e. iterators that do *not* have any loops and only
have a single top-level yield like `iter myIter() { yield 1; }`

This is not a common iterator idiom, and certainly not a performance critical
one. In fact there are only 3 tests that run into this, and none of them are
compelling.

This code isn't well understood by anybody on the team, and doesn't do anything
for performance, so it's being removed to simply iterator lowering.

https://github.com/chapel-lang/chapel/issues/5357 has some more info as well as
examples.