Add primitive array memory NUMA localization.
This adds a primitive capability to set the NUMA locality of array
memory.  In the runtime we add a chpl-topo module that can set NUMA
locality for a block of memory, either localizing the whole thing to a
single sublocale (NUMA domain) or doing a blocked localization across
all the sublocales.  DefaultRectangularArr:initialize() specifies in its
calls to _ddata_allocate() how to set the locality for array memory when
there are multiple sublocales, and ChapelBase:_ddata_allocate() calls
the new chpl_topo runtime code to do the actual work of localizing the
memory.