Test updates - 8 new tests and 2 modifications
Modified base-class-generic-args to stop relying on the old constructor strategy
of creating a generic instantiation from the arguments of the constructor.  Now
the test explicitly sets the generic fields during Phase 1.

I made phase1/omittedCastEtc a future temporarily, so that I wouldn't have to
worry about details like if a cast was needed right away.

Add 8 new tests, specifically for generics.

uninit-type.chpl was split off of generic-uninit-field-initializers.chpl so that
I would have a test that would exercise the part I was focused on (type fields),
instead of muddying its correct operation with the various parts exercised in
the original.  To put it plainly, uninit-type.chpl focuses on a class with a
single generic field which is a type field, and uses that type field as the type
for another field.  It only sets the value of the type field in its Phase 1 and
relies on the insertion of omitted field initialization for the second field.

init-type.chpl is a variant on uninit-type.chpl, which gives the generic type a
default value, but explicitly sets it during Phase1 of the initializer body.

init-type2.chpl is a variant on init-type.chpl.  This version does not take an
argument for the generic field with a default value - instead, it relies on the
insertion of omitted initialization to handle the setting of the generic field.
I had an error in my branch which ignored initializers on generic types when the
initializer itself did not have any other generic arguments, so this test
verifies that has been fixed.

init-type3.chpl is a variant on init-type2.chpl.  Instead of having no
arguments, this version has one argument, but that argument is not generic.  The
error which impacted init-type2.chpl also impacted this version, so I kept both
around.

init-type4.chpl is a variant on init-type3.chpl.  This time, we don't have an
argument for the type field, but our other argument is generic so the candidate
was not rejected (even when the error affecting init-type2.chpl and
init-type3.chpl was still present).

ordinary_param.chpl has a simple type with only a param field, and exercises the
many different ways to determine the type of the variable (even if the object
itself is not initialized).  This test is a future until I fully handle param
fields

omittedGenerics1.chpl exercises a case for params that was not already handled
by base-class-generic-args.chpl.  Specifically, it handles params that only
supply either a type or an initialization expression, but not both.  These
fields are omitted during the user's Phase 1, to exercise that code path.  This
test is a future until I fully handle param fields.

omittedGenerics2.chpl exercises a different case for params that was not already
handled by base-class-generic-args.chpl.  Specifically, it handles when a param
field supplies neither a type nor an initialization, and is still omitted in
Phase 1 of the body.  As expected, this test generates an error message due to
omitting the initialization of a field with no value we can substitute.