Clean up the init and exprType of field defPoints at the end of resolution
The code in the init and exprType at the defPoint of a field might contain
references to param and type fields.  Those fields (and any other type or param
variable) are removed at the end of resolution, but resolution doesn't expect
any init or exprType to be left on a defPoint, so it doesn't know to clean up
references to them there.  This change just blanket-removes all the init and
exprType values for fields at the end of resolution, for simplicity.  It assumes
that we make no new generic instantiations for a type after the completion of
function resolution.  Hopefully that assumption holds.