Modify some checks to no longer apply to initializers
checks.cpp was requiring that all AggregateTypes define a
defaultTypeConstructor.  My change for generic initializers no longer relies
on a type constructor at all, so make this check not apply

Change a check in functionResolution's resolveExpr to only try to operate on
the defaultTypeConstructor if one exists.  Prior to this change, generic
instantiations would run afoul of this section because they do not create a
default type constructor.  They also don't need any of the operations that were
provided by that section, so skipping them in that case is safe.

There is more work needed to remove the dependence on type constructors in
general, but this is a good waypoint towards that eventual goal.