Add test of Chapel array API and improve API somewhat
In working on array views, I decided to write a test that tried
to exercise all standard user-facing APIs on arrays in order to
ensure that they work on array views as well.  In doing so, I
made some improvements to the array API to make some cases work
better:

* added an error overload to IRV so that when called on non-sparse
  arrays it would give a nice message

* changed the where clauses on the array-vector ops into param
  conditionals with compilerError() messages to give errors when
  applied to arrays other than 1D default rectangulars.

* fixed a bug in the find() routine in which it would only work for 1D
  arrays (when I think we want it to work for general arrays, and
  nothing about its signature or documentation suggested otherwise)

This test is independent of the array views work, as are the
improvements, so I'm filing it against master separately.