Add support for a new primitive, PRIM_INITIALIZER_SET_FIELD, and use it
The new primitive is intended to support the creation of the generic
instantiation based on Phase 1 of an initializer for a generic aggregate type.
This strategy is a significant departure from how generic instantiations are
created for constructors, which relies on the generic fields having a
corresponding argument in every constructor argument list.  The utilization of
this primitive during function resolution allows the instantiation to be built
as Phase 1 of the body is traversed and resolved, ensuring its completion when
the call to the parent or other initializer has returned control to Phase 2 of
the current initializer.

Details:

checkAst.cpp:
checks.cpp:
primitive.h:
primitive.cpp:
optimizeOnClauses.cpp:
  Add basic handling of the new primitive

symbol.h:
symbol.cpp:
  Extend TypeSymbol's renaming upon instantiation to have a mode that is
compatible with the strategy our new initializers will use.  This mode accepts a
single update at a time, as there will only be one update per use of the new
primitive PRIM_INITIALIZER_SET_FIELD

type.h:
type.cpp:
  Add support for creating the new instantiation.  This centers around two
methods on AggregateTypes, getInstantiation and setNextGenericField.

setNextGenericField determines the location of the first generic field in this
AggregateType that does not have a substitution.  It does this by traversing all
the fields until it finds a type, param, or generic var/const field, and setting
its genericField index to that location, or zero if no generic fields were
found.

getInstantiation organizes generic AggregateTypes and their instantiations into
a tree of most generic to each instantiation for the current generic field until
we reach an AggregateType with no remaining generic fields (which is a leaf).
It does this by having each AggregateType store a vector of instantiations for
the current generic field, which themselves store a vector of instantiations for
the next generic field, etc.  This allows us to reuse partial instantiations
when resolving multiple initializers for the same type, if they have a common
set of values for the earliest generic fields.

getInstantiation first checks if the index being updated is earlier than our
first known unhandled generic field, in which case the type returns itself.
Then, if the index matches, it checks its list of known instantiations for the
current generic field.  If we find an instantiation that matches the value we
are expecting to put into this generic field, just return that instead of making
a new one.  If not, we need to create a new (potentially partial) instantiation
with this value.  We then put our substitution in and update the field
appropriately in the new instantiation, and determine the next generic field for
the instantiation, if any.  The instantiation is returned after this point.

initializerRules.cpp:
  During the verification that an initializer follows the rules we have
described, we make any initialization of a field during Phase 1 use the new
primitive PRIM_INITIALIZER_SET_FIELD.  This includes transforming user inserted
field initialization (which I created a new function to handle) and inserting
omitted initialization in this form.

functionResolution.cpp:
  Create a new function, resolveFieldInit, which is called when the new
primitive is encountered.  This function ensures we're making a valid field
access, and then calls the AggregateType method to get the instantiation for
this particular field update if the field is generic.  It replaces all
references to the type and its fields in the initializer with references to the
instantiated type.  When it is done, it transforms the new primitive into
PRIM_SET_MEMBER, as it should be treated identically after this point.