Let hwloc auto-detect libnuma for non-flat locModels; don't force it.
This fixes a flaw in #5418 (yesterday).  There I arranged to throw
--enable-libnuma to the hwloc configure step for locale models other
than 'flat'.  But when libnuma is not present on the system, doing so
will cause the configure step to produce an error.  On systems without
libnuma, we want to be able to use such locale models even though they
might be hobbled by the lack of NUMA info.  So instead of insisting that
hwloc use libnuma in this case, just let it auto-detect libnuma, and
we'll get by as best we can if it's not there.